+++ lightyear/src/shared/replication/hierarchy.rs
@@ -4,6 +4,7 @@ use bevy::ecs::entity::MapEntities;
 use bevy::prelude::*;
 use serde::{Deserialize, Serialize};
 
+use crate::prelude::client::{InterpolationSet, PredictionSet};
 use crate::prelude::server::ControlledBy;
 use crate::prelude::{
     MainSet, NetworkRelevanceMode, PrePredicted, Replicated, Replicating, ReplicationGroup,
@@ -259,6 +260,10 @@ impl<R: ReplicationPeer> Plugin for HierarchyReceivePlugin<R> {
             PreUpdate,
             Self::update_parent
                 .after(InternalMainSet::<R::SetMarker>::Receive)
+                // we want update_parent to run in the same frame that ParentSync is propagated
+                // to the predicted/interpolated entities
+                .after(PredictionSet::SpawnHistory)
+                .after(InterpolationSet::SpawnHistory)
                 // NOTE: we're putting this in MainSet::Receive so that users can order
                 // their systems after this
                 .in_set(MainSet::Receive),